fix audio transcript snippet resolution
