+++ archive/python-sdk-legacy/authentication.mdx
@@ -5,11 +5,11 @@ icon: "key"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: c00a82133845848c5ccba9c88cf02be7d3d05461f99237e6af3cab0645263637 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-authentication" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-authentication")} />
 </Visibility>
