+++ archive/python-sdk-legacy/index.mdx
@@ -3,11 +3,11 @@ title: "Legacy"
 description: "Archived documentation for the legacy Session-based Python SDK"
 icon: "clock-rotate-left"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: f8182e3340dca27d44a7ca1dda5055d137e23f36f5ae05cd4b9d2b909aff4e21 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-index" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-index")} />
 </Visibility>
