+++ archive/python-sdk-legacy/migration-guide.mdx
@@ -3,11 +3,11 @@ title: "Migration Guide"
 description: "Migrate from the legacy Session-based API to the modern client-based Fish Audio Python SDK"
 icon: "arrow-right-arrow-left"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: db9ee9309b347cf4484c3cdb97e88cbe272b534224d102a060da51950ac8c05a */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-migration-guide" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-migration-guide")} />
 </Visibility>
