+++ archive/python-sdk-legacy/text-to-speech.mdx
@@ -5,11 +5,11 @@ icon: "microphone"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 2e151e18b761b42c6158472c1f9e4b1729420c723f8c41f96e23a87f5b442951 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-text-to-speech" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-text-to-speech")} />
 </Visibility>
