+++ archive/python-sdk-legacy/voice-cloning.mdx
@@ -5,11 +5,11 @@ icon: "clone"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: df4ea781a900d8c8f6605b8782598ffba3879dfea424692438113630bae5d91c */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-voice-cloning" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-voice-cloning")} />
 </Visibility>
