+++ archive/python-sdk-legacy/websocket.mdx
@@ -5,11 +5,11 @@ icon: "bolt"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 18b25b81e58fb71294481f82aeb9f3c668fc1f28c68134d65820d41696041bc7 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-sdk-legacy-websocket" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-sdk-legacy-websocket")} />
 </Visibility>
