+++ developer-guide/best-practices/emotion-control.mdx
@@ -3,11 +3,11 @@ title: "Emotion & Expression Control"
 description: "Make your AI voices express emotions naturally"
 icon: "face-smile"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: da55415ae4c2483d1de3a286d0b7e74e1fe2c849dcd7de9742bfec9e0ad9e547 */}
 <Visibility for="humans">
-<PageAudioTranscript page="best-practices-emotion-control" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("best-practices-emotion-control")} />
 </Visibility>
