+++ developer-guide/core-features/fine-grained-control.mdx
@@ -4,11 +4,11 @@ description: 'Advanced control over speech generation'
 icon: "sliders"
 iconType: "solid"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 4a46ae86b04c62730f1554051768c47b306f2378a3624545d53de421d5d19acd */}
 <Visibility for="humans">
-<PageAudioTranscript page="core-features-fine-grained-control" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("core-features-fine-grained-control")} />
 </Visibility>
