+++ developer-guide/core-features/speech-to-text.mdx
@@ -3,11 +3,11 @@ title: "Speech to Text Guide"
 description: "Convert audio recordings into accurate text transcriptions"
 icon: "microphone-lines"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: d76297f22aba43ae28b2b18a4102cd1dd58a82a461670bbb30192a9195ae5bda */}
 <Visibility for="humans">
-<PageAudioTranscript page="core-features-speech-to-text" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("core-features-speech-to-text")} />
 </Visibility>
