+++ developer-guide/getting-started/quickstart.mdx
@@ -6,11 +6,11 @@ icon: "rocket"
 
 import Support from '/snippets/support.mdx';
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 38d7da4971b5246339f9a56ad260706a63218a91edff54b740ef249f36f9e7a2 */}
 <Visibility for="humans">
-<PageAudioTranscript page="getting-started-quickstart" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("getting-started-quickstart")} />
 </Visibility>
