+++ developer-guide/models-pricing/choosing-a-model.mdx
@@ -5,11 +5,11 @@ icon: "magnifying-glass"
 ---
 
 import Support from '/snippets/support.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 9ceeae8d692c3471e65687cc8770a2d3d31b5554c6e9ac6dfd77a1f477e716c1 */}
 <Visibility for="humans">
-<PageAudioTranscript page="models-pricing-choosing-a-model" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("models-pricing-choosing-a-model")} />
 </Visibility>
