+++ developer-guide/models-pricing/models-overview.mdx
@@ -5,11 +5,11 @@ icon: "layer-group"
 ---
 
 import Support from '/snippets/support.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 4395e380edbecf49d53e7f3467921d76195bec8400ac22b5a37c2b3c52613a29 */}
 <Visibility for="humans">
-<PageAudioTranscript page="models-pricing-models-overview" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("models-pricing-models-overview")} />
 </Visibility>
