+++ developer-guide/products/story-studio.mdx
@@ -3,11 +3,11 @@ title: "Story Studio"
 description: "Build immersive audio stories and narratives"
 icon: "wand-magic-sparkles"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: de306c9f3b406de604d8bad8fabe70e7d955d04a136d1aaf49c69d0ea87dd072 */}
 <Visibility for="humans">
-<PageAudioTranscript page="products-story-studio" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("products-story-studio")} />
 </Visibility>
