+++ developer-guide/resources/coding-agents.mdx
@@ -3,11 +3,11 @@ title: "AI Coding Agents"
 description: "Connect AI coding assistants to Fish Audio documentation via MCP for real-time API guidance"
 icon: "robot"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: bb48040abec73604a76c532d042b645b95eeccb2d2d82929888bf41b3180935a */}
 <Visibility for="humans">
-<PageAudioTranscript page="resources-coding-agents" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("resources-coding-agents")} />
 </Visibility>
