+++ developer-guide/resources/migration.mdx
@@ -3,11 +3,11 @@ title: "Migration Guide"
 description: "Switch from ElevenLabs, OpenAI, or other TTS providers to Fish Audio"
 icon: "arrow-right-arrow-left"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: c215f39e8e8f5d56a6dd2f4f770673e67e8392e0cdda36a70906fbbad83c9957 */}
 <Visibility for="humans">
-<PageAudioTranscript page="resources-migration" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("resources-migration")} />
 </Visibility>
