+++ developer-guide/sdk-guide/javascript/authentication.mdx
@@ -5,11 +5,11 @@ icon: "key"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 8dffb4526a7b17be3092274000e01772b599c37b7307a4f8404cb633dfca7c72 */}
 <Visibility for="humans">
-<PageAudioTranscript page="javascript-authentication" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("javascript-authentication")} />
 </Visibility>
