+++ developer-guide/sdk-guide/javascript/text-to-speech.mdx
@@ -5,11 +5,11 @@ icon: "microphone"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: fc17eef01159e7d16739b323c5a330a562a917a11e29489e3ccfc436fe8c5841 */}
 <Visibility for="humans">
-<PageAudioTranscript page="javascript-text-to-speech" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("javascript-text-to-speech")} />
 </Visibility>
