+++ developer-guide/sdk-guide/python/overview.mdx
@@ -5,11 +5,11 @@ icon: "python"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 76979ffd115d3a77f60fd200d42e49605cce2cc21a06c8f380cfafff92443525 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-overview" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-overview")} />
 </Visibility>
