+++ developer-guide/sdk-guide/python/speech-to-text.mdx
@@ -5,11 +5,11 @@ icon: "microphone-lines"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 7f58d0e3f5a776bec6496f58879f50737b57f27bc2e0fb54e0cd37598df71ac8 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-speech-to-text" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-speech-to-text")} />
 </Visibility>
