+++ developer-guide/sdk-guide/python/text-to-speech.mdx
@@ -5,11 +5,11 @@ icon: "microphone"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 8449d7e9db3c781cf36b33bb4913e5b883353cb2fde48a7c5ac1670fbbd6feb5 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-text-to-speech" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-text-to-speech")} />
 </Visibility>
