+++ developer-guide/sdk-guide/python/websocket.mdx
@@ -5,11 +5,11 @@ icon: "bolt"
 ---
 
 import Prerequisites from '/snippets/prerequisites.mdx';
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: e1fba201bbf1e887ba6b4783a3435cbdf78ebf9094f5ba29ac495725f19f09c7 */}
 <Visibility for="humans">
-<PageAudioTranscript page="python-websocket" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("python-websocket")} />
 </Visibility>
