+++ developer-guide/self-hosting/docker-deployment.mdx
@@ -3,11 +3,11 @@ title: "Docker Deployment"
 description: "Deploy Fish Audio models using Docker containers"
 icon: "docker"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: 7519d2e10765e9a8af2318d43268e2c1f0fe2db4478dadeec9afa3068c0dcb98 */}
 <Visibility for="humans">
-<PageAudioTranscript page="self-hosting-docker-deployment" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("self-hosting-docker-deployment")} />
 </Visibility>
