+++ developer-guide/self-hosting/local-setup.mdx
@@ -3,11 +3,11 @@ title: "Local Model Setup"
 description: "Install and configure Fish Audio models for local inference"
 icon: "server"
 ---
-import { PageAudioTranscript } from '/snippets/page-audio-transcript.jsx';
+import { AudioTranscript, getPageAudioTranscriptVoices } from '/snippets/audio-transcript.jsx';
 
 {/* speak-mintlify-hash: f1038df54a82e77cfe449b4182e5a43b9739fce33ecc1347496c5db4efd98e09 */}
 <Visibility for="humans">
-<PageAudioTranscript page="self-hosting-local-setup" />
+  <AudioTranscript voices={getPageAudioTranscriptVoices("self-hosting-local-setup")} />
 </Visibility>
